refactor: standardizes the controllers of endpoints `/migrations` and `/status`
