+++ infra/database.js
@@ -1,4 +1,5 @@
 import { Client } from "pg";
+import { ServiceError } from "./errors";
 
 async function query(queryObject) {
   let client;
@@ -7,9 +8,12 @@ async function query(queryObject) {
     const result = await client.query(queryObject);
     return result;
   } catch (error) {
-    console.log("Error dentro do databse.js");
-    console.error(error);
-    throw error;
+    const serviceErrorObject = new ServiceError({
+      message: "Erro na conexão ou Banco ou na Query.",
+      cause: error,
+    });
+    console.error(serviceErrorObject);
+    throw serviceErrorObject;
   } finally {
     await client?.end();
   }
